+++ test.base/test/java/test/base/TestProvider1.java
@@ -12,6 +12,21 @@ public class TestProvider1 implements ToolProvider {
 
   @Override
   public int run(PrintWriter out, PrintWriter err, String... args) {
-    return 0;
+    try {
+      assertThatAssertionsAreEnabled();
+      return 0;
+    } catch (Throwable throwable) {
+      throwable.printStackTrace(err);
+      return 1;
+    }
+  }
+
+  private void assertThatAssertionsAreEnabled() {
+    try {
+      assert false;
+    } catch (AssertionError expected) {
+      return;
+    }
+    throw new AssertionError("Assertions are not enabled for " + getClass().getClassLoader());
   }
 }